ACTIN-3199: Filter ageGroup adult/senior in algo
